+++ src/test/java/gov/ca/cwds/idm/ResendInvitationEmailIdmResourceTest.java
@@ -1,0 +1,48 @@
+package gov.ca.cwds.idm;
+
+import static gov.ca.cwds.config.api.idm.Roles.OFFICE_ADMIN;
+import static gov.ca.cwds.config.api.idm.Roles.STATE_ADMIN;
+
+import org.junit.Test;
+import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
+import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
+
+public class ResendInvitationEmailIdmResourceTest extends IdmResourceTest {
+
+  @Test
+  @WithMockCustomUser(county = "OtherCounty")
+  public void testResendInvitationEmailWithDifferentCounty() throws Exception {
+    assertResendEmailUnauthorized(YOLO_COUNTY_USERS_EMAIL);
+  }
+
+  @Test
+  @WithMockCustomUser(roles = {OFFICE_ADMIN}, adminOfficeIds = {"otherOfficeId"})
+  public void testResendInvitationEmailWithOfficeRole() throws Exception {
+    assertResendEmailUnauthorized(YOLO_COUNTY_USERS_EMAIL);
+  }
+
+  @Test
+  @WithMockCustomUser(roles = {"OtherRole"})
+  public void testResendInvitationEmailWithOtherRole() throws Exception {
+    assertResendEmailUnauthorized(YOLO_COUNTY_USERS_EMAIL);
+  }
+
+  @Test
+  @WithMockCustomUser(roles = {STATE_ADMIN}, county = "Madera")
+  public void testResendInvitationEmailWithStateAdmin() throws Exception {
+    assertResendEmailWorksFine();
+  }
+
+  @Test
+  @WithMockCustomUser()
+  public void testResendInvitationEmailWithCountyAdmin() throws Exception {
+    assertResendEmailWorksFine();
+  }
+
+  private void assertResendEmailUnauthorized(String email) throws Exception {
+    mockMvc
+        .perform(MockMvcRequestBuilders.get("/idm/users/resend?email=" + email))
+        .andExpect(MockMvcResultMatchers.status().isUnauthorized())
+        .andReturn();
+  }
+}
