+++ src/test/java/gov/ca/cwds/idm/SecondIdmResourceTest.java
@@ -61,41 +61,4 @@ public class SecondIdmResourceTest extends IdmResourceTest {
 
     assertCreateUserBadRequest(user, "fixtures/idm/create-user/no-racfid-in-cws-error.json");
   }
-
-  @Test
-  @WithMockCustomUser(county = "OtherCounty")
-  public void testResendInvitationEmailWithDifferentCounty() throws Exception {
-    assertResendEmailUnauthorized(YOLO_COUNTY_USERS_EMAIL);
-  }
-
-  @Test
-  @WithMockCustomUser(roles = {OFFICE_ADMIN}, adminOfficeIds = {"otherOfficeId"})
-  public void testResendInvitationEmailWithOfficeRole() throws Exception {
-    assertResendEmailUnauthorized(YOLO_COUNTY_USERS_EMAIL);
-  }
-
-  @Test
-  @WithMockCustomUser(roles = {"OtherRole"})
-  public void testResendInvitationEmailWithOtherRole() throws Exception {
-    assertResendEmailUnauthorized(YOLO_COUNTY_USERS_EMAIL);
-  }
-
-  private void assertResendEmailUnauthorized(String email) throws Exception {
-    mockMvc
-        .perform(MockMvcRequestBuilders.get("/idm/users/resend?email=" + email))
-        .andExpect(MockMvcResultMatchers.status().isUnauthorized())
-        .andReturn();
-  }
-
-  @Test
-  @WithMockCustomUser(roles = {STATE_ADMIN}, county = "Madera")
-  public void testResendInvitationEmailWithStateAdmin() throws Exception {
-    assertResendEmailWorksFine();
-  }
-
-  @Test
-  @WithMockCustomUser()
-  public void testResendInvitationEmailWithCountyAdmin() throws Exception {
-    assertResendEmailWorksFine();
-  }
 }
